+++ Tests/Functional/Service/JsonlRoundtripTest.php
@@ -48,10 +48,9 @@ class JsonlRoundtripTest extends FunctionalTestCase
 
         $this->get(ImportService::class)->runImport($file, 0, ['workspaceId' => 0]);
 
-        $qb = GeneralUtility::makeInstance(ConnectionPool::class)->getQueryBuilderForTable('pages');
-        $qb->getRestrictions()->removeAll();
-        $imported = (int)$qb->count('uid')->from('pages')
-            ->where($qb->expr()->gt('tx_impexpnl_remote_uid', $qb->createNamedParameter(0)))
+        $qb = GeneralUtility::makeInstance(ConnectionPool::class)->getQueryBuilderForTable('tx_impexpnl_uid_map');
+        $imported = (int)$qb->count('uid')->from('tx_impexpnl_uid_map')
+            ->where($qb->expr()->eq('table_name', $qb->createNamedParameter('pages')))
             ->executeQuery()->fetchOne();
 
         self::assertGreaterThan(0, $imported, 'JSONL-Import hat keine Seiten angelegt');